requirements.txt
Flask==3.0.3
pywebview==5.3
reportlab==4.2.5
python-ulid==3.0.0
